fix: Migrate cache to Redis Module
