+++ src/app/app.module.ts
@@ -17,11 +17,11 @@ import { BankAccountModule } from '../modules/bank_account/bank_account.module';
 import constantConfig from '../config/constant.config';
 import mailConfig from '../config/mail_service.config';
 import { MailModule } from '../modules/mail/mail.module';
-import { AuthModule } from 'src/modules/auth/auth.module';
 import redisConfig from 'src/config/redis.config';
-import { SetCacheBlockedUserUsecase } from 'src/modules/auth/core/usecases';
-import Redis from 'ioredis';
 import { BankConfigModule } from 'src/modules/bank_config/bank_config.module';
+import { AuthModule } from 'src/modules/auth/auth.module';
+import { SetCacheBlockedUserUsecase } from 'src/modules/redis_cache/core/usecases';
+import { RedisCacheModule } from 'src/modules/redis_cache/redis_cache.module';
 
 @Module({
   imports: [
@@ -50,31 +50,17 @@ import { BankConfigModule } from 'src/modules/bank_config/bank_config.module';
         return addTransactionalDataSource(new DataSource(options));
       },
     }),
+    forwardRef(() => RedisCacheModule),
     forwardRef(() => UserModule),
     forwardRef(() => AuthModule),
     forwardRef(() => BankModule),
     forwardRef(() => BankAccountModule),
     forwardRef(() => MailModule),
-    forwardRef(() => AuthModule),
     forwardRef(() => BankConfigModule),
   ],
   controllers: [AppController],
-  providers: [
-    {
-      provide: 'REDIS_CLIENT',
-      useFactory: async (configService: ConfigService) => {
-        const redisConfig = configService.get('redis');
-        return new Redis(redisConfig);
-      },
-      inject: [ConfigService],
-    },
-    {
-      provide: 'CircularDependencyLog',
-      useFactory: () => (msg: string) =>
-        console.warn(`Circular dependency: ${msg}`),
-    },
-  ],
-  exports: ['REDIS_CLIENT'],
+  providers: [],
+  exports: [],
 })
 export class AppModule implements OnModuleInit {
   constructor(
