+++ src/modules/mail/core/models/forgot_password_data.model.ts
