+++ src/modules/mail/core/models/get_initial_password_data.model.ts
