+++ src/modules/redis_cache/core/usecases/block_user/set_cache_blocked_user.usecase.ts
@@ -1,11 +1,11 @@
 import { forwardRef, Inject, Injectable } from '@nestjs/common';
-import { CacheBlockedUserIRepo } from '../../repositories/cache_blocked_user.irepo';
 import { GetBlockedUserUsecase } from 'src/modules/user/core/usecases';
+import { ICacheBlockedUserRepo } from '../../repositories/cache_blocked_user.irepo';
 
 @Injectable()
 export class SetCacheBlockedUserUsecase {
   constructor(
-    private readonly cacheBlockedUserRepo: CacheBlockedUserIRepo,
+    private readonly cacheBlockedUserRepo: ICacheBlockedUserRepo,
 
     @Inject(forwardRef(() => GetBlockedUserUsecase))
     private readonly getBlockedUserUsecase: GetBlockedUserUsecase,
